+++ packages/aws-client/utils.js
@@ -1,0 +1,54 @@
+const pRetry = require('p-retry');
+const { setErrorStack } = require('@cumulus/common/util');
+
+/**
+ * Wrap a function and provide a better stack trace
+ *
+ * If a call is made to the aws-sdk and it causes an exception, the stack trace
+ * that is returned gives no indication of where the error actually occurred.
+ *
+ * This utility will wrap a function and, when it is called, update any raised
+ * error with a better stack trace.
+ *
+ * @param {Function} fn - the function to wrap
+ * @returns {Function} a wrapper function
+ */
+const improveStackTrace = (fn) =>
+  async (...args) => {
+    const tracerError = {};
+    try {
+      Error.captureStackTrace(tracerError);
+      return await fn(...args);
+    } catch (err) {
+      setErrorStack(err, tracerError.stack);
+      err.message = `${err.message}; Function params: ${JSON.stringify(args, null, 2)}`;
+      throw err;
+    }
+  };
+
+const retryIfThrottlingException = (err) => {
+  if (exports.isThrottlingException(err)) throw err;
+  throw new pRetry.AbortError(err);
+};
+
+/**
+ * Wrap a function so that it will retry when a ThrottlingException is encountered.
+ *
+ * @param {Function} fn - the function to retry.  This function must return a Promise.
+ * @param {Object} options - retry options, documented here:
+ *   - https://github.com/sindresorhus/p-retry#options
+ *   - https://github.com/tim-kos/node-retry#retryoperationoptions
+ *   - https://github.com/tim-kos/node-retry#retrytimeoutsoptions
+ * @returns {Function} a function that will retry on a ThrottlingException
+ */
+const retryOnThrottlingException = (fn, options) =>
+  (...args) =>
+    pRetry(
+      () => fn(...args).catch(retryIfThrottlingException),
+      { maxTimeout: 5000, ...options }
+    );
+
+module.exports = {
+  improveStackTrace,
+  retryOnThrottlingException
+};
